requirements.txt
streamlit==1.22.0
python-dotenv==0.19.2
requests==2.28.1